guis/pyqt5/lcbci_lab_pyqt5/pyproject.toml
[tool.poetry]
name = "lcbci_lab"
version = "0.1.0"
description = ""
authors = ["chanhakim <chanhakim17@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyqtgraph = "^0.11.0"
pyserial = "^3.4"
pyqt5 = "^5.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
lcbci = "lcbci_lab._app:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
